Changed the position/length properties in the retrieval service to be longs instead of ints, and fixed Retrieval's getPercentage() method

+++ plugins/au.gov.ga.earthsci.core/src/au/gov/ga/earthsci/core/retrieve/RetrieverMonitorDelegate.java
@@ -36,19 +36,19 @@ public class RetrieverMonitorDelegate implements IRetrieverMonitor
 	}
 
 	@Override
-	public void progress(int amount)
+	public void progress(long amount)
 	{
 		delegate.progress(amount);
 	}
 
 	@Override
-	public void setPosition(int position)
+	public void setPosition(long position)
 	{
 		delegate.setPosition(position);
 	}
 
 	@Override
-	public void setLength(int length)
+	public void setLength(long length)
 	{
 		delegate.setLength(length);
 	}
